feat: add recharging

+++ common/src/main/java/com/daqem/irobot/item/data/BatteryDataComponent.java
@@ -5,22 +5,22 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
 
-public record BatteryDataComponent(int energy, int maxEnergy) {
+public record BatteryDataComponent(double energy, double maxEnergy) {
 
     public static final Codec<BatteryDataComponent> CODEC = RecordCodecBuilder.create(instance -> instance.group(
-            Codec.INT.fieldOf("energy").forGetter(BatteryDataComponent::energy),
-            Codec.INT.fieldOf("max_energy").forGetter(BatteryDataComponent::maxEnergy)
+            Codec.DOUBLE.fieldOf("energy").forGetter(BatteryDataComponent::energy),
+            Codec.DOUBLE.fieldOf("max_energy").forGetter(BatteryDataComponent::maxEnergy)
     ).apply(instance, BatteryDataComponent::new));
 
     public static final StreamCodec<RegistryFriendlyByteBuf, BatteryDataComponent> STREAM_CODEC = StreamCodec.of(
             (buf, data) -> {
-                buf.writeInt(data.energy());
-                buf.writeInt(data.maxEnergy());
+                buf.writeDouble(data.energy());
+                buf.writeDouble(data.maxEnergy());
             },
-            buf -> new BatteryDataComponent(buf.readInt(), buf.readInt())
+            buf -> new BatteryDataComponent(buf.readDouble(), buf.readDouble())
     );
 
-    public BatteryDataComponent withEnergy(int energy) {
+    public BatteryDataComponent withEnergy(double energy) {
         return new BatteryDataComponent(Math.max(0, Math.min(energy, this.maxEnergy)), this.maxEnergy);
     }
 }
